adicionado o emptypersitent para substituir a abstração
